LIBFCREPO-469. Moved argparse config for commands into modules.
Subcommands are now Command classes in the command.{cmd_name} package. An instance of each command is created upon starting plastron, where the configuration runs.

https://issues.umd.edu/browse/LIBFCREPO-469
