+++ commands/ping.py
@@ -1,0 +1,13 @@
+from classes.exceptions import FailureException
+
+class Command:
+    def __init__(self, subparsers):
+        parser_ping = subparsers.add_parser('ping',
+                description='Check connection to the repository')
+        parser_ping.set_defaults(cmd_name='ping')
+
+    def __call__(self, fcrepo, args):
+        try:
+            fcrepo.test_connection()
+        except:
+            raise FailureException()
